Merge QSPI and QSPI RAMDISK targets
Merge the mentioned targets where possible, and make the rest coherent.

Make sure the script name ends in _ramdisk when it is used for booting
ramdisks.
Earlier, such suffix was used only if there were two scripts. We want
the naming to be coherent, even if there is only one boot script for a
target.